cvlib: give context in case of renderer fail
Note that we convert the item's stringification from Unicode to bytes. If this
isn't done, the exception text may not be printed in normal usage. This is
because in normal usage the output of this program is piped into a file, in
which case Python tried to convert any output from Unicode to bytes with the
ascii codec. (This is true even though usually stdout is piped and stderr goes
to a terminal.) If there are non-ASCII characters in the item (not unlikely),
it breaks and the result is to suppress the exception text.