Add a "refpreprint" category for publications.
This should be set to "y" if the publication is not yet refereed, but is a
preprint of something that is expected to appear in a refereed venue. If
you're (say) applying for jobs and want to show all of your awesome
publications, then it is worthwhile to show these. It's not honest to label
them as refereed, but putting them in a "non-refereed" section generally
obscures both their impact and their freshness. I list these as "preprints"
personally. This seems like a fair way to go.