worklog.py: add a multi-line BEGIN_SUBST command
This is intended to provide a more convenient means of inserting useful
computed quantities straight into a document. The CITESTATS command attempted
to do this but I relied on a separate little text file to provide the
substitution template, which is definitely a bit annoying. Here I just added a
very quick mechanism for commands that take multiple template lines as input.