wlgithub.py: tidy up GitHub querying
There is more to do here. The BigQuery calls are getting big enough that
they will incur charges with Google, and there's no need to redundantly
query back to 2011 every time we do an update.

Also, when I ran the script more-or-less as-is, it took 10 hours and
eventually crashed. But then I ran the same queries in the GCloud
BigQuery console and they all finished in just a few minutes. What's the
deal with that??? Need to figure that out too.